All intro pages finished. See description for more details
Finished the intro pages (8).
The goal is now to let the user sign in with all their data that they just selected throughout the intro pages. The user needs to sign up using:
-username
-email
-password

The moment the user is signed up, their data should also automatically be saved to their account in the User collection on the website Appwrite.